Fix camera realignment, add SwimAboveWater state
Fix camera realignment after return from a non-kinematic state (outside the animator), where the fish has been unparented, then moved using physics. The fishParent is moved to the correct location relative to the fish, then the parent relationship between fish and fishParent is reestablished.
Add a state where the fish is kinematic, but swims above the surface of the water. The state transitions in this case from the Swim state to the SwimAboveWater state. In the SwimAboveWater state, the fish is switched to a non-kinematic state (outside the animator), and therefore falls back into the water due to gravity.